frontend: Move useTypedSelector to redux/hooks and decouple
This reduces coupling between the rest of the app and redux.

Because the useTypedSelector only needs to import a type it
does not import all of the reducers.

This means there's less import cycles in the app.
