+++ frontend/src/components/App/Notifications/Notifications.tsx
@@ -34,7 +34,7 @@ import { FixedSizeList, ListChildComponentProps } from 'react-window';
 import { useClustersConf } from '../../../lib/k8s';
 import Event from '../../../lib/k8s/event';
 import { createRouteURL } from '../../../lib/router';
-import { useTypedSelector } from '../../../redux/reducers/reducers';
+import { useTypedSelector } from '../../../redux/hooks';
 import Empty from '../../common/EmptyContent';
 import { DateLabel } from '../../common/Label';
 import {
