+++ frontend/src/components/App/Settings/DrawerModeSettings.tsx
@@ -20,7 +20,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import { setDetailDrawerEnabled } from '../../../redux/drawerModeSlice';
-import { useTypedSelector } from '../../../redux/reducers/reducers';
+import { useTypedSelector } from '../../../redux/hooks';
 import { TooltipIcon } from '../../common/Tooltip';
 
 export default function DrawerModeSettings() {
