+++ frontend/src/components/App/VersionDialog.tsx
@@ -18,7 +18,7 @@ import DialogContent from '@mui/material/DialogContent';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import { getProductName, getVersion } from '../../helpers/getProductInfo';
-import { useTypedSelector } from '../../redux/reducers/reducers';
+import { useTypedSelector } from '../../redux/hooks';
 import { uiSlice } from '../../redux/uiSlice';
 import { Dialog } from '../common/Dialog';
 import NameValueTable from '../common/NameValueTable';
