+++ frontend/src/components/Sidebar/NavigationTabs.tsx
@@ -23,7 +23,7 @@ import { useTranslation } from 'react-i18next';
 import { generatePath, useHistory } from 'react-router';
 import { getCluster, getClusterPrefixedPath } from '../../lib/cluster';
 import { createRouteURL } from '../../lib/router';
-import { useTypedSelector } from '../../redux/reducers/reducers';
+import { useTypedSelector } from '../../redux/hooks';
 import Tabs from '../common/Tabs';
 import { SidebarItemProps } from '../Sidebar';
 import { useSidebarItems } from './useSidebarItems';
