+++ frontend/src/components/common/ActionsNotifier.tsx
@@ -21,7 +21,7 @@ import React, { useRef } from 'react';
 import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { CLUSTER_ACTION_GRACE_PERIOD, ClusterAction } from '../../redux/clusterActionSlice';
-import { useTypedSelector } from '../../redux/reducers/reducers';
+import { useTypedSelector } from '../../redux/hooks';
 
 export interface PureActionsNotifierProps {
   clusterActions: { [x: string]: ClusterAction };
