+++ frontend/src/components/resourceMap/kubeIcon/KubeIcon.tsx
@@ -16,7 +16,7 @@
 
 import Box from '@mui/material/Box';
 import { alpha } from '@mui/system/colorManipulator';
-import { useTypedSelector } from '../../../redux/reducers/reducers';
+import { useTypedSelector } from '../../../redux/hooks';
 import CRoleIcon from './img/c-role.svg?react';
 import CmIcon from './img/cm.svg?react';
 import CrbIcon from './img/crb.svg?react';
