+++ frontend/src/plugin/Plugins.tsx
@@ -21,7 +21,7 @@ import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { isElectron } from '../helpers/isElectron';
-import { useTypedSelector } from '../redux/reducers/reducers';
+import { useTypedSelector } from '../redux/hooks';
 import { fetchAndExecutePlugins } from './index';
 import { pluginsLoaded, setPluginSettings } from './pluginsSlice';
