+++ frontend/src/redux/hooks.ts
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-import { useDispatch, useSelector, useStore } from 'react-redux';
-import { AppDispatch, AppStore, RootState } from './stores/store';
+import { TypedUseSelectorHook, useDispatch, useSelector, useStore } from 'react-redux';
+import type { AppDispatch, AppStore, RootState } from './stores/store';
 
 // These know the types from our store.
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
 export const useAppSelector = useSelector.withTypes<RootState>();
 export const useAppStore = useStore.withTypes<AppStore>();
+export const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector;
