+++ frontend/src/redux/reducers/reducers.tsx
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import { TypedUseSelectorHook, useSelector } from 'react-redux';
 import { combineReducers } from 'redux';
 import notificationsReducer from '../../components/App/Notifications/notificationsSlice';
 import themeReducer from '../../components/App/themeSlice';
@@ -58,7 +57,5 @@ const reducers = combineReducers({
 });
 
 export type RootState = ReturnType<typeof reducers>;
-
-export const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector;
 
 export default reducers;
